task 0. print array with func as arg

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,4 @@
 #include "function_pointers.h"
-
 /**
  * array_iterator - executes function with array element
  * @array: array of element to iterate and pass to func
@@ -10,7 +9,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array != null && action != null)
 		for (i = 0; i < size; i++)
